refactor : [ #5 의존성 역전 ]
