+++ MemberService/src/main/java/com/memberservice/member/domain/UUIDFactory.java
@@ -1,0 +1,16 @@
+package com.memberservice.member.domain;
+
+import com.memberservice.member.service.port.IdentifierFactory;
+import org.springframework.stereotype.Component;
+
+import java.util.UUID;
+
+@Component
+public class UUIDFactory implements IdentifierFactory {
+
+
+    @Override
+    public String generate() {
+        return UUID.randomUUID().toString();
+    }
+}
