+++ MemberService/src/main/java/com/memberservice/member/domain/dto/response/ResponseMemberGatewayDto.java
@@ -1,4 +1,4 @@
-package com.memberservice.member.model.dto.response;
+package com.memberservice.member.domain.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
