--- MemberService/src/main/java/com/memberservice/member/repository/MemberRepository.java
@@ -1,15 +1,0 @@
-package com.memberservice.member.repository;
-
-import com.memberservice.member.model.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByEmail(String email);
-
-    Optional<Member> findByMemberId(String memberId);
-}
