+++ MemberService/src/main/java/com/memberservice/member/service/port/IdentifierFactory.java
@@ -1,0 +1,7 @@
+package com.memberservice.member.service.port;
+
+public interface IdentifierFactory {
+
+    String generate();
+
+}
