+++ MemberService/src/main/java/com/memberservice/member/service/port/MemberRepository.java
@@ -1,0 +1,16 @@
+package com.memberservice.member.service.port;
+
+import com.memberservice.member.domain.entity.Member;
+
+import java.util.Optional;
+
+public interface MemberRepository {
+
+    Member save(Member member);
+
+    Optional<Member> findById(Long id);
+
+    Optional<Member> findByEmail(String email);
+
+    Optional<Member> findByMemberId(String memberId);
+}
