+++ MemberService/src/test/java/com/memberservice/member/service/MemberServiceTest.java
@@ -1,0 +1,29 @@
+//package com.memberservice.member.service;
+//
+//import com.fasterxml.jackson.databind.ObjectMapper;
+//import org.junit.jupiter.api.BeforeEach;
+//import org.junit.jupiter.api.DisplayName;
+//import org.junit.jupiter.api.Test;
+//import org.springframework.security.crypto.password.PasswordEncoder;
+//
+//class MemberServiceTest {
+//
+//    private MemberService memberService;
+//
+//    @BeforeEach
+//    void init() {
+//        this.memberService = new MemberService(
+//
+//        );
+//    }
+//
+//    private MemberRepository memberRepository;
+//    private PasswordEncoder passwordEncoder;
+//    private ObjectMapper objectMapper;
+//    private IdentifierFactory identifierFactory;
+//
+//    @Test
+//    @DisplayName("회원가입")
+//    void signUp() {
+//    }
+//}
